update routerlicious deps on common-utils (#6884)
Part ??? of fixing #6757

Going through bumping all the layers to use prerelease versions of common-utils following a change there. This change bumps usages in serve/routerlicious.

Next steps:

Create prerelease for server 0.1028
Update client and tinylicious to use common-utils 0.32.0-0 and server 0.1028.0-0. Tinylicious must be updated here because the common-utils change is hashing changes to facilitate local testing in insecure environments, and that entire dependency chain needs to be updated.