Fixing potential issues around client having connection that does not broadcast ops (#6928)
We observed that some clients in ODSP are not observing any ops on a "write" connection for an hour, until token expired. On reconnect, client finds itself being behind with many thousands of ops.

We do not have good theory, but I went through client code looking for cases where we might either not have proper "op" handler installed, or maybe connection is completely broken. Based on code inspection:

Due to asynchrony, DeltaManager might get closed connection as a new connection and not realize that. Currently there is no way to check for that state, so adding IDisposable to connection to be able to detect it and recover.
event registration logic is a bit complicated and while code inspection did not find a bug, I want to be able to assert that event registration / propagation is done properly, so refactoring this code and adding more asserts.

More details are can be found in https://onedrive.visualstudio.com/SPIN/_queries/edit/1156694/?triage=true
Also related PR: https://github.com/microsoft/FluidFramework/pull/6931