Pull out opsUntilFirstConnect check and use opsSinceLastAck instead (#6907)
Instead of doing "join" op sequence number minus DeltaManager.initialSequenceNumber, we use SummaryCollection.opsSinceLastAck. The former was a count of ops we "caught up" with, but the latter is ops since the last summary ack. Both should be equivalent when loading from latest snapshot, but in cases of loading from cached snapshot, the latter is more accurate.

Also removes use of PromiseTimer for initial delay, and instead uses the new delay common-utils function.
Also pulls out the handler to further reduce number of ContainerContext references within SummaryManager.

Changed the logic to just check opsSinceLastAck at the point of deciding whether to initial delay or not. Now it uses
a new checkBypassInitialDelay() function in conjunction with a deferred to simplify this logic. We also now check more frequently if we need to bypass the delay- any time refreshSummarizer() is called in either Off or Starting state (Starting State is new here).