Fixing a bug: always fetching ops from PUSH (#7017)
Fetching ops from push was under "if (from < this.firstCacheMiss)" check that resulted in not requesting ops from PUSH all the time.
Fixing it by making sure fetching from PUSH is unique callback that is always involved.

How found:

In case of customDimensions.containerId == "65f88376-9665-465b-bc79-ae18d0ea0647", forcing reconnect actually resolved an issue of stalled client, because initial ops resolved the op gap. Client was stalled prior to that for two reasons:

fetching ops was first not bringing any ops, and then it started timing out.
The fact that it was resolved on reconnect forced me to inspect code, and I see that we do not always request ops from PUSH when we are looking for ops. That's the bug