(split from Whiteboard) Merged PR 33126: fix: prevent superfluous edit in reconciliation path
The logic for computing the reconciliation path sometimes added a superfluous edit (i.e. an edit known to the issuer of the edit being applied at the time the edit applied was issued) at the start of the path.

This had no negative impact aside from a minor perf impact.

Related work items: #61287