(split from Whiteboard) Merged PR 33172: refactor: BREAKING CHANGE: clarify some of the intermediate stages of document to domain
This:
- removes an unneeded overload from documentRootToDomainNodePolymorphic and does related cleanup
- adds and uses `ValidatedNode` type alias
- starts unifying arguments for projecting of validated nodes into ProjectionContext
- call into more specific parts of the projection logic from the reencoding logic to avoid redundant validation and loss of context.
- project intermediate domain trees with editing disabled (since its unused).

This is a breaking change since it makes some small changes to the public API for options for document to domain (ex: exported type names for options changes). No behavioral changes should be needed, and most code should work as is.

This change is making progress toward clear separation and modularization of validation vs conversion/projection, focused on passing the right data to re encoding and back to document to domain.

Related work items: #62086