Replace CreateContainerError usages with either normalizeError or new GenericError (#6940)
CreateContainerError as-is is basically equivalent to normalizeError*. So wherever CreateContainerError was given a thrown object, just use normalizeError.

We also called CreateContainerError to raise new cases, those should directly create GenericError with the string used as the error code as well as message.

I also reimplemented CreateProcessingError with normalizeError in here but could split into a different PR.

* Here are the differences between what CreateContainerError returned, and what normalizeError now returns where it's used instead:

The returned error is no longer an instance of LoggingError. So arbitrary props set on there won't be logged, you must use addTelemetryProperties
"Partially" valid errors - e.g. with an errorType but no telemetry prop functions or vice-versa - will not have those properties brought over. These are hypothetical cases that don't happen in practice so we decided to stop supporting them for simplicity.
Tangentially related to @vladsud 's PR #6936