We may miss "disconnected" event on socket (#6986)
When connection is established, connection object is returned back to DeltaManager. During this transition time, nobody listens on "disconnected" event, make it possible to miss it.

Add proper handling for such cases, as well as validation that if connection object it not disposed, socket itself should be connected.

We had someone similar behavior earlier for "error" handler, but it was not fully correct. Extending it to "disconnect" event and fixing issues.