More reliable last Summary flow (#7247)
Rework logic around managing

The primary motivation for this change is to ensure that last summary always starts and has a chance to run in cases where document has too many ops (i.e. 10K limit in ODSP). Together with PR #7230 they sort of try to pull in different directions - PR #7230 ensures that summarization process cancels as quickly as possible when summarizer client loses connection, while this PR ensures that when main client decides it needs to move to exit, it gives a chance to summarizer client (if it is still connected) to run through full summary. It also ensures that new summarizer (interactive) client has a delay before it starts summarizer instance, thus reducing chances of collisions with another (running last summary) summarizer. Together, they substantially reduce number of cases where we see summary nacks due to overlapping summarizers. It does not eliminate it completely, because RunningSummarizer.trySummarize does 3 attempts to summarize, and it needs to try only once for lastSummary (this is subject for another change).

Most of the changes are reworking logic in SummaryManager as I found it hard to follow when logic is split between multiple functions when they call each other in various ways. Now main flow is one function - SummaryManager.startSummarization(), which I find much easier to follow. Plus I've added a bunch of asserts to ensure that state transition are easier to follow (and we catch any issues).

One core change here is that there was a delay before staring a summarizer client, but it only applied to when container was booting (i.e. if summarizer client was created at that moment). There was no delay when given client (that was already running for a while) was assigned summarization role - it would start summarizer right away, increasing substantially odds of stepping on previous summarizer client that is doing last summary.

I believe there were some other smaller issues, like not re-checking if given client is still elected summarizer after various async processes, leading to more chances that we have two summarizers running in parallel.

Long term, we should move away from delays as a way to solve these problems and look into ability for current summarizing (non-interactive) client to be elected as summarizer, such that last summary that keeps running for 10 mins never has to compete with other clients (but that's subject for another design / PR).

There are certain small bits & pieces that are spill over from #7230 (i.e. they resolve to nothing after that PR), like stop() reason in couple places. These PRs can go in any order, but it's better for this go last to create cleaner diff.