Elected summarizer client occasionally does not start summarizer. (#7321)
Close #7311: Summarizer election may result in summarizer not being started
Much more details are in https://onedrive.visualstudio.com/SPIN/_queries/edit/1186576/?triage=true

The core of the problem that client who is assigned to be a summarizer never creates summarizing client.
So I've been doing a bunch of code reviewing to figure out all possible reasons for why it may happen.
But looking back at history for this container, it actually was elected summarizer before, and summarizer started, but never finished and there were no events regarding the end. So that pointed to it never finishing booting in the first place.
So from code, it looks like a condition that was never handled correctly - container being disposed before summarizer even had a chance to start. So addressing this.

Also while reviewing all the ways how we propagate events RE leader election, we might run into state where events and state is not synchronized due to an early return statement in summarizerClientElection.ts. Fixing that too.