Type Validation: Specify Breaking Changes in Package.json (#7353)
This change modifies the format of the tests by removing variables, and making the back and forward compact portions independent. In addition, it adds the ability to specify breaking changes in the package.json directly.

Note. this change does not introduce any breaks. it simply adds entries for existing breaks