add more detached blob tests (#7325)
Add some more tests for detached blob support.

Add .tmp extension for empty files. This was removed previously but is re added here. The extension is expected to be removed by the host, which is not tested here.
Add a blob to serialized summary which will cause rehydration to fail if no detached blob storage is provided.