Change no join op recovery to telemetry only (#7392)
Issue #7312: fluid:telemetry:DeltaManager:ConnectionRecovery is hit in production and ODSP scalability tests

I've been going around this area for quite a while, both trying to better understand it, but also try various recovery options.
Unfortunately none of the recovery options tried so far does not really work. So removing it and switching to only logging more data to keep understanding problem better.

With that, it's worth also provide a bit more info on what I've learned so far: Big part of these issues, at least in stress tests, is the fact that we boot container with "write" connection, and connection is established before container is loaded. As result, because load may take a while, we are not processing ops for very long time.
I'll attempt to solve this in separate PR, but adding DeltaManager.handler & Container.loaded properties to payload is super valuable, as well as learning that we do sit on ton of unprocessed ops.