Take advantage of detached state for initializing model (#7351)
Since FluidContainer supports detached state now, we can initialize the data prior to attaching and drop the initialize call which waits on the key to be ready. This is recommendable and less error-prone (it's easy to forget that the second client might arrive before a post-attach value is set). Here I do this via a static method on the controller.

This change also evicts some view logic from the app and restructures the view a bit -- probably this should just be converted to React at this point but I stop short of that here.