Better telemetry and recovery for cases where Container does not catch up and does not realize that  (#7430)
Please see issue #7312 for more details.
This change does two things:
1. It adds telemetry to detect cases where we know that we are behind, but we are not actively trying to catch up. We do not fully understand these problems, but this telemetry (along with other PRs that add telemetry data points in various places) will help us better understand the problem.
2. It also engages recovery for such cases.

We know for sure how we can get into such cases with "read" connections. These are normal (expected) workflows.
"write" connections are much more mysterious, as service (PUSH for ODSP) guarantees that join op should not be missed and should be delivered to client quickly after connection. But telemetry shows that we may be in position where join op is not processed for minutes, and it's not clear at the moment if that's a bug on client or server.

Semi-related PRs that together helps us move in right direction here:
PR #7428
PR #7427
PR #7429