Include more packages in TSDocs/API docs build (#7472)
The TSDocs are currently exempt from the broken link checker because there were so many broken links. The reason is that there are some packages that aren't included in the TSDocs currently, leading to broken links. Long-term, these packages should be refactored, but for now, this PR just adds all the needed packages to the docs build so that the links are fixed.