Add docs page for breaking notes (#7479)
This is to prepare for the more detailed changelogs that we expect to be publishing in the future after public preview.

Some notes:

I added a section in the docs site as a placeholder for changelogs. The expectation is that the section's index page will link to all the change notes pages, where each release gets its own page (image of how it looks attached)
The existing BREAKING.md will be reviewed at release for formatting etc. and a release notes page for that release will be created under docs/changelogs/releases/<some release identifier>, and a link to it will be added under the main changelogs index. This is currently manual, and future work is expected to automate adding additional information, such as released version numbers or additional individual change information
This change doesn't address any of the automation stuff yet (soon™)