(split from Whiteboard) Merged PR 34520: docs: Fix casing of `@inheritDoc` comment tags
VSCode intellisense does not respect `@inheritdoc`; it only supports `@inheritDoc`.

This PR updates comments in the code to use the latter form of the tag.

I have filed an issue with tsdoc to either fix support for the all lower-case form, or update the eslint plugin to not allow it: https://github.com/microsoft/tsdoc/issues/301