Roll up re-exported APIs in docs pipeline (#7497)
api-extractor doesn't pull TSDocs from external re-exports. This means that the TSDocs for the uber fluid-framework package is empty

This change adds a script to manually merge the members from the api-extractor JSON files for the 'source' packages into a fluid-framework.api.json. This isn't a 'correct' fix, rather a pragmatic one.