Remove OpProcessingController, use LoaderContainerTracker in propertyDDS tests (#7784)
Slight changes in reconnect logic (see PRs #7753, #7393) result in failures in PropertyDDS UTs.

Looking a bit deeper, I can't easily follow intentions in OpProcessingController.
For example, OpProcessingController.process(dm1) will leave all but dm1 paused. UTs do rely on that (it's pretty clearly from some of them), in other places it feels like it's unintended result. And I do not think it correctly waits for all pending activity to be flushed (as observed in above mentioned PRs).

Given that it's deprecated and repo uses LoaderContainerTracker (explicitly or implicitly through), it's time to do the move.

Some random tests fail with this change, but are addressed by increasing timeout.
We saw this problem before with these tests - see pending #7350