DeltaManager must not throw at disconnect when it still has messages in the buffer (#7645)
While doing some testing in the Bohemia repo to enable removing the batch manager and setting the default flush mode in the runtime to TurnBased (see #7365) I ran into a situation in which the DeltaManager gets disconnected with messages in its buffer. This leads to an error being thrown which ends with the container closing. However, if the container just reconnects, the pending messages are being re-sent.