Review of _cancel events: change some to be an error, vs. others not to be an error. (#7841)
Change some to be an error, vs. others not to be an error.
Simplify code in couple places.

Overall theme - most of the events should be FYI kind of events. The top level scenario event is the one that should report an error.
I.e. in case of container attachment, we report an error only through container close event, though most errors will also be reported through containerAttach_cancel (as that's where most errors are).
None of low-level events report issues as errors.

Similar, only ObtainSnapshot event reports errors, not any sub-steps.

Closes #7804 - please see it for more details