Type Validation: Detect Removed Types (#7871)
Previously we just skipped types that existed in the old package, but not the new package, which will miss breaks. Now we will try to use the type even though it doesn't exist in the new package. There are valid cases where a type can change, these tests will now validate those, and for things that are truly removed, those breaks can be captured in the package json like any other