Type Validator: Ignore Pre-Release (#7881)
In the ci build we always bump the version in the package.json to a pre-release version. This creates an issue where we try to bump the type validation version when the version has not been released. This change normalizes the version to remove the pre-release, and will only bump the validation version if there is a change in that.

fixes #7875

related to #7614