Remove internal usage of IFluidObject for  `IFluidModule`, and `IContainerContext.scope` (#7867)
This change is in preparation for some changes coming to eventually remove IFluidObject as an uber interface, as the coupling of all types off a single interface leads to complex typing issues. This change just moves away from IFluidObject on some types that come from the definitions package, IFluidModule, and IContainerContext.scope. In a follow up i will be removing IFluidObject entirely from these types

related to #7613