start of adding progress reporting & cancelation / no warnings on storage calls (#7837)
We leverage container warnings today to communicate to host whenever we have any problems with connectivity.
I believe this is wrong (my mistake!) and the right way to deal with issues is by communicating status directly to caller (i.e. each call site is in full control and is responsible to figure out if any messaging is required to user).

Warnings have a lot of issues:

- There is "on" signal (i.e. start of throttling), but no "off" event (or rather we do not communicate to host how to pair them, and in many cases there are no "off" events).
- In most cases error / progress UX should be local. I.e. insert image scenario should show in-place progress / error UX, not use global container banner to tell the user that there are some issues with image upload. This composes much more naturally - if user deletes such pending upload, there is no more UX (even if call is not cancelled right away), whereas today we will keep showing error banner.
This change is just starting point. It does not change any public API yet, but it does remove warnings from telemetry.
See a bit more on how it shows up in telemetry here: #7804

In the future we should
- ensure that every call has cancellation token.
- There is proper handling of delays in our stack. For example, summary upload logic may decided to abandon upload if it did not happen for the long time and regenerate summary using newer state (before trying to upload again).