send op when joining with pending/stashed ops (#7811)
Send an op containing previous client ID when reconnecting with pending ops or when connecting for the first time with stashed ops. This fixes the current limitation of only being able to get stashed ops from a container when closing it by allowing future containers to recognize when the original container reconnects, or a new container is loaded with the stashed ops, and the ops are then successfully submitted with a new client ID that differs from the one in the pending state. This will allow hosts to checkpoint the pending state of a container and have a better ability to prevent data loss in the case of unexpected container closure.