(split from Whiteboard) Merged PR 35599: refactor: BREAKING CHANGE: Separate public and internal change types
Changes submitted to a SharedTree are currently given as-is directly to the fluid service. This PR adds a translation step that converts public, user-created changes into internal changes that are submitted to the service. This:
1. No longer requires the user to know the "wire-format" of changes; they can use a potentially simpler public API instead.
2. Gives SharedTree an opportunity to translate the public changes into a more optimal internal format, e.g. to optimize the change format for performance before sending it.
3. Prevents a user from being able to pass extraneous data to the fluid service, e.g. objects that satisfy the API but which have many additional properties that would be unnecessarily serialized to the network.
4. Prevents a user from passing changes to the SharedTree but then mutating them later (SharedTree now does a deep defensive copy of all changes).

The conversion code is located in overrides of the method `GenericSharedTree.internalizeChanges()`. Currently all implementations simply clone the changes as they are - no optimizations or format changes are made yet.

In practice, this means that SharedTrees are now parameterized over both an external/public `TChange` and an internal `TChangeInternal`. `Checkout` therefore must also be parameterized over both. Most other classes that are parameterized over TChange (e.g. EditLog, Transaction) now use the internal change type rather than the public one. The concrete public/internal types in use now are Change/ChangeInternal and AnchoredChange/AnchoredChangeInternal for SharedTree and SharedTreeWithAnchors, respectively. In the future, the hope is that the public change types might be replaced by more ergonomic editing APIs.

Several other minor refactors occurred as part of this work as well:
* OrderedEditSet no longer provides an API to summarize (EditLog does).
* SharedTree and Checkout now have `revert()` methods, rather than users reverting the changes manually themselves.
* A helper and type has been added for isomorphic transformations of TreeNode trees (this is currently employed when internalizing Build changes)

## Downstream Integration Notes
This PR breaks one usages of `Edit<Change>` in whiteboard-app code - a PR to sidestep that [is here](https://dev.azure.com/intentional/intent/_git/whiteboard-app/pullrequest/35678).

Related work items: #66549