Remove offline summarizer errors from error table (#8064)
Client becoming offline at different times of summarizer lifetime are prevailing reason for errors in telemetry.
These are not really errors, i.e. these are expected states that we should not bring attention of OCEs / developers, unless it keeps happening for long time (which is covered by such events as "Behind").

So this is my attempt to downplay severity and stop reporting them as errors.
It's a bit ugly, and in general goes against the rule of doing proper transformation through queries (and keeping telemetry "raw"). At the same time, we really want our telemetry errors to represents real issues that developers and OCEs need to pay attention, and keep the messaging simple.