Add new code details API surface to Container and remove chaincodePackage (#8104)
This is the first step of adding the new code details API surface to migrate to. It adds the methods as optional on the interface (since interop with prior versions that don't have it is required), as well as the implementation.

Since chaincodePackage has been deprecated for over a year and is already absent from the IContainer interface, this also removes that API. The codeDetails property is also now marked deprecated for removal in a future release.