Record how long it takes BatchManager to complete full batch (#8232)
Make processing of ops async, the same way it is in DeltaQueue.
This helps with UTs actually testing what we ship, but also (due to to small change in order in BatchManager) is now asserted by BatchManager, i.e. no reentrancy is allowed by inbound queue being synchronous.
Some renames in UT due to this separation.

And the main change - add error if it takes us more than 5 seconds to get full batch. The number is arbitrary, and I we will likely change it in future (as we will hit these errors in normal flow, which means they should not be errors). But for now, this is in sync with "OpRoundtripTime" errors that report latency delays as errors at 5 second mark, and the primary reason for this change.

Closes #8214
Helps with continued investigation of #7406
Natural continuation of PR #8206