Use require.resolve to determine paths to Webpack loaders (#8297)
Package hoisting with tools like yarn makes it impossible to be sure that the layout of the node_modules folder will always be the same. In fact, depending on the exact install strategy, the node_modules folders may not even exist. More info at https://yarnpkg.com/advanced/rulebook#modules-shouldnt-hardcode-node_modules-paths-to-access-other-modules.