Quick fix for the current issue with updating snapshots
Currently, generating snapshots hangs. However, the mocha test harness does not.

I still believe there is value in investigating exactly why the tool is failing (#8335) as it may uncover a more serious issue with either the present tool or the replay tool or the container lifecycle itself. But at the same time, it would be great to incorporate this process in the test pipeline to at least have it running more often - not included in this PR.