Move tsd tests to normal build (#8349)
This change removes our dependency on tsd, which wasn't easily integrated into the build, and didn't work with things like linting. Now the those tests are just like our typeValidation tests and build normally with our build, and leverage @ts-expect-error instead of the tsd function expectError, which was really the only thing we needed.

Additionally, tsd was quite a heavy dependency, so should reduce a tiny amount of our install time