port type validation fix to main (#8374)
The latest version of protocol def uses namespaces, which leads to type names with periods . in them. There is a bug in the test generation that keeps these periods in the function names, which is not valid syntax. This change simple replaces ' with _ in the function names.

Once this is in, I'll release the build tools, bump the version, and re-add generation in protocol defs

related to #8373