On-demand summarization using a summarizer client created on the fly (#8251)
Support the case in which a summarizer and client are created on the fly and used to perform on-demand summarization. Related API's: Summarizer.create is a static method that creates a Summarizer and underlying client. Summarizer.summarizeOnDemand launches a summary using the summarizer produced by the create() method. Summarizer.stop() halts and disposes the summarizer and closes the summarizer client.