Deprecate SharedNumberSequence and SharedObjectSequence, and remove most usage of SharedObjectSequence (#8585)
This change is part of #8526

This change marks SharedNumberSequence and SharedObjectSequence as deprecated to discourage new usage, and removes most SharedObjectSequence usage in our repository (replay-tool is left unchanged). It also removes the useSyncedArray from @fluid-experimental/react which depended on SharedObjectSequence.

Changes to examples:

likes-and-comments: Removed the comments portion which relied on SharedObjectSequence, leaving it as just "likes"
badge: Removed history functionality
shared-text: Removed test object
monaco: Removed extraneous registry entry
todo: Convert back to map-based approach
I recognize that some of these example changes are not ideal, but I think we should continue discussion of the scenario in #8526.